+++ cobib/commands/list.py
@@ -122,7 +122,7 @@ class ListCommand(Command):
                 pass
             tui.list_args += ['-s']
         # handle input via prompt
-        command, _ = tui.prompt_handler('list ' + ' '.join(tui.list_args), out=tui.buffer)
+        command, _ = tui.execute_command('list ' + ' '.join(tui.list_args), out=tui.buffer)
         # after the command has been executed n the prompt handler, the `command` variable will
         # contain the contents of the prompt
         LOGGER.debug('Post-process ListCommand arguments for consistent prompt.')