+++ cobib/commands/open.py
@@ -5,6 +5,7 @@ import logging
 import os
 import subprocess
 import sys
+from collections import defaultdict
 from urllib.parse import urlparse
 
 from cobib.config import CONFIG
@@ -18,6 +19,7 @@ class OpenCommand(Command):
 
     name = 'open'
 
+    # pylint: disable=too-many-branches
     def execute(self, args, out=sys.stderr):
         """Open file from entries.
 
@@ -37,40 +39,102 @@ class OpenCommand(Command):
             largs = parser.parse_args(args)
         except argparse.ArgumentError as exc:
             print("{}: {}".format(exc.argument_name, exc.message), file=sys.stderr)
-            return None
+            return
 
-        errors = []
+        # pylint: disable=too-many-nested-blocks
         for label in largs.labels:
+            things_to_open = defaultdict(list)
+            count = 0
+            # first: find all possible things to open
             try:
                 entry = CONFIG.config['BIB_DATA'][label]
-                if 'file' not in entry.data.keys() or entry.data['file'] is None:
-                    msg = "Error: There is no file associated with this entry."
-                    LOGGER.error(msg)
-                    if out is None:
-                        errors.append(msg)
-                    continue
-                opener = CONFIG.config['DATABASE'].get('open', None)
-                try:
-                    LOGGER.debug('Parsing "%s" for URLs.', entry.data['file'])
-                    url = urlparse(entry.data['file'])
-                    if url.scheme:
-                        # actual URL
-                        url = url.geturl()
-                    else:
-                        # assume we are talking about a file and thus get its absolute path
-                        url = os.path.abspath(url.geturl())
-                    LOGGER.debug('Opening "%s" with %s.', url, opener)
-                    with open(os.devnull, 'w') as devnull:
-                        subprocess.Popen([opener, url], stdout=devnull, stderr=devnull,
-                                         stdin=devnull, close_fds=True)
-                except FileNotFoundError as err:
-                    LOGGER.error(err)
-                    errors.append(str(err))
+                for field in ('file', 'url'):
+                    if field in entry.data.keys() and entry.data[field]:
+                        for val in entry.data[field].split(','):
+                            val = val.strip()
+                            LOGGER.debug('Parsing "%s" for URLs.', val)
+                            things_to_open[field] += [urlparse(val)]
+                            count += 1
             except KeyError:
                 msg = "Error: No entry with the label '{}' could be found.".format(label)
                 LOGGER.error(msg)
+                continue
 
-        return '\n'.join(errors)
+            # if there are none, skip current label
+            if not things_to_open:
+                msg = "Warning: This entry has no actionable field associated with it."
+                LOGGER.warning(msg)
+                print(msg, file=out or sys.stderr)
+                continue
+
+            if count == 1:
+                # we found a single URL to open
+                self._open_url(list(things_to_open.values())[0][0])
+            else:
+                # we query the user what to do
+                idx = 1
+                url_list = []
+                prompt = []
+                # print formatted list of available URLs
+                for field, urls in things_to_open.items():
+                    for url in urls:
+                        prompt.append(f"{idx:3}: [{field}] {url.geturl()}")
+                        url_list.append(url)
+                        idx += 1
+                # loop until the user picks a valid choice
+                help_requested = False
+                while True:
+                    prompt_copy = prompt.copy()
+                    prompt_copy.append("Entry to open [Type 'help' for more info]: ")
+                    try:
+                        choice = input('\n'.join(prompt_copy)).strip()
+                    except EOFError:
+                        choice = ''
+                    if not choice:
+                        # empty input
+                        msg = 'User aborted open command.'
+                        LOGGER.warning(msg)
+                        print(msg, file=sys.stderr)
+                        break
+                    if choice == 'help':
+                        LOGGER.debug('User requested help.')
+                        if not help_requested:
+                            msg = ["You can specify one of the following options:",
+                                   "  1. a url number",
+                                   "  2. a field name provided in '[...]'",
+                                   "  3. or simply 'all'",
+                                   "  4. ENTER will abort the command",
+                                   ""]
+                            prompt = msg + prompt
+                        help_requested = True
+                    elif choice == 'all':
+                        LOGGER.debug('User selected all urls.')
+                        for url in url_list:
+                            self._open_url(url)
+                        break
+                    elif choice in things_to_open.keys():
+                        LOGGER.debug('User selected the %s set of urls.', choice)
+                        for url in things_to_open[choice]:
+                            self._open_url(url)
+                        break
+                    elif choice.isdigit() and int(choice) > 0 and int(choice) <= count:
+                        LOGGER.debug('User selected url %s', choice)
+                        self._open_url(url_list[int(choice)-1])
+                        break
+
+    @staticmethod
+    def _open_url(url):
+        """Opens a URL."""
+        opener = CONFIG.config['DATABASE'].get('open', None)
+        try:
+            url = url.geturl() if url.scheme else os.path.abspath(url.geturl())
+            LOGGER.debug('Opening "%s" with %s.', url, opener)
+            with open(os.devnull, 'w') as devnull:
+                subprocess.Popen([opener, url], stdout=devnull, stderr=devnull,
+                                 stdin=devnull, close_fds=True)
+        except FileNotFoundError as err:
+            LOGGER.error(err)
+            print(err, file=sys.stderr)
 
     @staticmethod
     def tui(tui):
@@ -83,7 +147,4 @@ class OpenCommand(Command):
             # get current label
             label, _ = tui.get_current_label()
             labels = [label]
-        # populate buffer with entry data
-        error = OpenCommand().execute(labels, out=None)
-        if error:
-            tui.prompt_print(error)
+        tui.execute_command(['open'] + labels, skip_prompt=True)