+++ cobib/commands/show.py
@@ -61,7 +61,7 @@ class ShowCommand(Command):
             LOGGER.debug('Current entry is selected. Applying highlighting.')
             tui.buffer.replace(0, label, CONFIG.get_ansi_color('selection') + label + '\x1b[0m')
         LOGGER.debug('Populating buffer with ShowCommand result.')
-        tui.buffer.view(tui.viewport, tui.visible, tui.width-1, tui.ANSI_MAP)
+        tui.buffer.view(tui.viewport, tui.visible, tui.width-1, ansi_map=tui.ANSI_MAP)
 
         # reset current cursor position
         tui.top_line = 0