+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Pypi: https://pypi.org/project/cobib/2.4.1/
 
+### Added
+- support for multiple associated files (#42,!19)
+- interactive menu when opening an entry with multiple associated files (!19)
+
 ### Fixed
 - ISBN parser was missing the ENTRYTYPE and did not use strings for number fields
 - the ISBN parser can now handle empty entries (#53)